refactor(element): move all decorators to source root
