+++ packages/element/__tests__/computingPair.ts
@@ -1,5 +1,5 @@
 // tslint:disable:max-classes-per-file
-import {ComputingPair, createComputingPair} from '../../src';
+import {ComputingPair, createComputingPair} from '../src';
 
 const repeatGetTenTimes = <C extends object, M extends keyof C, T extends C[M]>(
   instance: C,
