+++ packages/element/src/index.js
@@ -1,6 +1,6 @@
-export {default as attribute} from './decorators/attribute';
-export {default as createComputingPair} from './decorators/computingPair';
-export {default as createElementDecorator} from './decorators/element';
-export {default as internal} from './decorators/internal';
-export {default as property} from './decorators/property';
+export {default as attribute} from './attribute';
+export {default as createComputingPair} from './computingPair';
+export {default as createElementDecorator} from './element';
+export {default as internal} from './internal';
+export {default as property} from './property';
 export * from './tokens/lifecycle';
